+++ packages/client/hmi-client/src/router/index.ts
@@ -9,7 +9,6 @@ import UserAdminView from '@/components/navbar/tera-useradmin.vue';
 
 // These are test/experiment pages
 import ResponsivePlayground from '@/temp/ResponsivePlayground.vue';
-import TheiaView from '@/temp/theia.vue';
 import SSE from '@/temp/sse.vue';
 import EvaluationScenarios from '@/temp/EvaluationScenarios.vue';
 import AMRPetriTest from '@/temp/AMRPetriTest.vue';
@@ -29,7 +28,6 @@ export enum RoutePath {
 	Unauthorized = '/unauthorized',
 
 	// Playground and experiments, these components are testing-only
-	Theia = '/theia',
 	Ta2Playground = '/ta2-playground',
 	ResponsivePlaygroundPath = '/responsive-playground',
 	EvaluationScenariosPath = '/evaluation-scenarios'
@@ -62,7 +60,6 @@ const routes = [
 	},
 
 	// Playground and experiments, these components are testing-only
-	{ path: RoutePath.Theia, component: TheiaView },
 	{ path: RoutePath.ResponsivePlaygroundPath, component: ResponsivePlayground },
 	{ path: RoutePath.EvaluationScenariosPath, component: EvaluationScenarios },
 	{ path: '/sse', component: SSE },
