+++ packages/client/hmi-client/src/services/calibrate-workflow.ts
@@ -1,5 +1,3 @@
-import * as d3 from 'd3';
-
 import type { Dataset, CsvAsset } from '@/types/Types';
 import { getModelConfigurationById, getObservables } from '@/services/model-configurations';
 import { downloadRawFile, getDataset } from '@/services/dataset';
@@ -65,100 +63,4 @@ export const setupDatasetInput = async (datasetId: string | undefined) => {
 		return {};
 	}
 	return {};
-};
-
-export const renderLossGraph = (element: HTMLElement, data: any[], options: { width: number; height: number }) => {
-	const marginTop = 16;
-	const marginBottom = 45;
-	const marginLeft = 70;
-	const marginRight = 20;
-
-	const { width, height } = options;
-	const elemSelection = d3.select(element);
-	let svg: any = elemSelection.select('svg');
-	if (svg.empty()) {
-		svg = elemSelection.append('svg').attr('width', width).attr('height', height);
-		svg.append('g').append('path').attr('class', 'line');
-	}
-	const group = svg.select('g');
-	const path = group.select('.line');
-
-	const [minX, maxX] = d3.extent(data, (d) => d.iter);
-	const [minY, maxY] = d3.extent(data, (d) => d.loss);
-
-	const xScale = d3
-		.scaleLinear()
-		.domain([minX, maxX])
-		.range([marginLeft, width - marginRight]);
-	const yScale = d3
-		.scaleLinear()
-		.domain([Math.min(minY, 0), maxY])
-		.range([height - marginBottom, marginTop]);
-
-	const pathFn = d3
-		.line()
-		.x((d: any) => xScale(d.iter))
-		.y((d: any) => yScale(d.loss))
-		.curve(d3.curveBasis);
-
-	// Update the data and path
-	path
-		.datum(data)
-		.attr('d', pathFn(data))
-		.style('stroke', '#1B8073')
-		.style('stroke-width', '2px')
-		.style('fill', 'none');
-
-	// Add x-axis
-	const xAxis = d3.axisBottom(xScale).ticks(5);
-	let xAxisGroup = svg.select('.x-axis');
-	if (xAxisGroup.empty()) {
-		xAxisGroup = svg.append('g').attr('class', 'x-axis');
-	}
-	xAxisGroup
-		.attr('transform', `translate(0, ${height - marginBottom})`)
-		.call(xAxis)
-		.selectAll('text')
-		.style('fill', '#6A7682')
-		.style('font-family', 'Figtree');
-	xAxisGroup.select('.domain').style('stroke', '#EAEAEA');
-	xAxisGroup.selectAll('.tick line').style('stroke', '#EAEAEA');
-	xAxisGroup
-		.append('text')
-		.attr('class', 'axis-label')
-		.attr('x', width / 2)
-		.attr('y', 35)
-		.style('text-anchor', 'middle')
-		.style('fill', '#101828')
-		.style('font-family', 'Figtree')
-		.style('font-size', '12')
-		.style('font-weight', '600')
-		.text('Solver iterations');
-
-	// Add y-axis
-	const yAxis = d3.axisLeft(yScale).ticks(3).tickFormat(d3.format('.1e'));
-	let yAxisGroup = svg.select('.y-axis');
-	if (yAxisGroup.empty()) {
-		yAxisGroup = svg.append('g').attr('class', 'y-axis');
-	}
-	yAxisGroup
-		.attr('transform', `translate(${marginLeft}, 0)`)
-		.call(yAxis)
-		.selectAll('text')
-		.style('fill', '#6A7682')
-		.style('font-family', 'Figtree');
-	yAxisGroup.select('.domain').style('stroke', '#EAEAEA');
-	yAxisGroup.selectAll('.tick line').style('stroke', '#EAEAEA');
-	yAxisGroup
-		.append('text')
-		.attr('class', 'axis-label')
-		.attr('transform', 'rotate(-90)')
-		.attr('y', -50) // Adjust this value to move the label within bounds
-		.attr('x', -(height / 2.5))
-		.style('text-anchor', 'middle')
-		.style('fill', '#101828')
-		.style('font-family', 'Figtree')
-		.style('font-size', '12')
-		.style('font-weight', '600')
-		.text('Loss');
 };
